requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
pandas==2.1.0
python-dotenv==1.0.0
huggingface-hub==0.24.0